Don't show border for 0 percent
Solves a graphical glitch where setting "color" to "white" on a "black" background and "bgColor" and "shadowColor" also set to "black" creates a very light but disturbing white border on the left, when percent is 0.  Obviously the calls to "renderHalfCircle" provoke that. So if percent is 0, just don't call "renderHalfCircle" as it is unnecessary anyway